requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy>=2.0.25
pytest==7.4.3
